multer error show as form error and updated logic to upload file and store db.

+++ controllers/authController.js
@@ -13,17 +13,18 @@ const profileUpdate = async (req, res) => {
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
-    await fs.unlinkSync(req.files[0].path);
+    req.files.length > 0 && (await fs.unlinkSync(req.files[0].path));
     return res.status(422).json(errorResource(errors.array(), 422));
   }
 
-  const validateData = matchedData(req);
   const user = req.user;
-  user.photo && (await fs.unlinkSync(`public/uploads/${user.photo}`));
-  await User.updateOne(
-    { _id: user._id },
-    { ...validateData, photo: req.files[0].filename }
-  );
+  let validateData = matchedData(req);
+  if (req.files.length > 0) {
+    user.photo && (await fs.unlinkSync(`public/uploads/${user.photo}`));
+    validateData.photo = req.files[0].filename;
+  }
+
+  await User.updateOne({ _id: user._id }, { ...validateData });
   const updatedUser = await User.findById(user.id).select({
     password: 0,
     __v: 0,
